EP11: Move EP11 MK change code into separate source file
No logic change, just moving existing code into a new source file
ep11_mkchange.c. This requires that certain definitions are moved into
ep11_specific.h, and some functions can no longer be static.
